Phase 1: Extract logging_setup.py from cli.py
Moves logging config resolution and log-level handling out of cli.py:
- resolve_logging_config (renamed from _resolve_logging_config — now
  the public API of its own module, no longer a private cli helper)
- add_log_level_arg
- apply_log_level

cli.py imports from the new module so existing call sites continue to
work. tests/unit/test_logging_config_resolution.py updated to import
from rushti.logging_setup with the new (un-prefixed) name.

Also drops now-unused configparser, Optional, and LOG_LEVELS imports
from cli.py.
